[wip] send more info to pool
